chore: revert rename of Token/Word to Word/Subword
